completing hpo
